+++ BugsnagPerformance/Assets/BugsnagPerformance/Scripts/Internal/PValueUpdater.cs
@@ -28,9 +28,9 @@ namespace BugsnagUnityPerformance
 
         public void Start()
         {
-            MainThreadDispatchBehaviour.Instance().Enqueue(CheckPValue());
+            MainThreadDispatchBehaviour.Enqueue(CheckPValue());
         }
-        
+
         private IEnumerator CheckPValue()
         {
 #if BUGSNAG_DEBUG